Update structure of PlantDetailPage
