--- frontend/src/Components/PlantDetailCard/PlantDetailCard.md
@@ -1,12 +1,0 @@
-```js
-<PlantDetailCard
-  plant={{
-    species: 'Monstera deliciosa',
-    info: 'variegated',
-    daysLeft: 4,
-    weeksLeft: 2,
-    waterInterval: 10,
-    fertilizerInterval: 4,
-  }}
-/>
-```
