+++ frontend/src/Components/PlantDetails/PlantDetails.md
@@ -1,0 +1,8 @@
+```js
+<PlantDetails
+  plant={{
+    species: 'Monstera deliciosa',
+    info: 'variegated',
+  }}
+/>
+```
